budtech-bees-automation-test-scripts/requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
atomicwrites==1.4.0
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.0.12
colorama==0.4.5
et-xmlfile==1.1.0
execnet==1.9.0
Faker==13.3.3
idna==3.3
iniconfig==1.1.1
jsonpath==0.82
mysql-connector==2.2.9
numpy==1.21.0
openpyxl==3.0.9
packaging==21.3
pandas==1.1.5
pluggy==1.0.0
py==1.11.0
pycryptodome==3.14.1
pyparsing==3.0.9
pytest==7.1.1
pytest-base-url==1.4.2
pytest-forked==1.4.0
pytest-html==3.1.1
pytest-metadata==2.0.1
pytest-ordering==0.6
pytest-rerunfailures==10.2
pytest-xdist==2.5.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.9
xlrd==1.2.0